debian: (Re-)adding private headers
Many people asked for it and even provided patches on the internet.
Interestingly there is not even a little interest to add packages for these headers, although there are patches ready to use.
I inspired myself on a patch that was made one year ago. However, I had to edit the rules and not-installed file to make it fit to the latest packaging files.

Credits: Helge Bahmann:
https://www.mail-archive.com/debian-qt-kde@lists.debian.org/msg88268.html

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com> (github: thopiekar)